style: Made style changes in the main file and modularized the header and search filter components.
