+++ src/components/Home/HomeJobFilter/JobFilter.tsx
@@ -1,0 +1,39 @@
+import { useNavigate } from 'react-router-dom';
+import { Form, FormWrapper, Input, SearchButton } from './styles';
+import { useState } from 'react';
+
+const JobFilter: React.FC<any> = ({ active }) => {
+    const [searchTerm, setSearchTerm] = useState('');
+    const navigate = useNavigate();
+    const handleSubmit = (event: any) => {
+        event.preventDefault();
+        navigate(`/feedvagas?searchTerm=${searchTerm}`);
+    };
+    return (
+        <FormWrapper active={active}>
+            <Form onSubmit={handleSubmit}>
+                <Input
+                    type="text"
+                    onChange={(event) => setSearchTerm(event.target.value)}
+                    active={active}
+                    placeholder="Digite o cargo, tipo de contrato ou localidade desejados"
+                />
+                <SearchButton
+                    type="submit"
+                    onClick={() =>
+                        window.scrollTo({
+                            top: 0,
+                            left: 0,
+                            behavior: 'smooth',
+                        })
+                    }
+                    active={active}
+                >
+                    Buscar Vaga
+                </SearchButton>
+            </Form>
+        </FormWrapper>
+    );
+};
+
+export default JobFilter;
